FIX134 - Rename the validation middlewares folder
